refactor: 이동 규칙을 MoveRule 인터페이스로 추상화
