+++ src/main/java/domain/move/rule/BasicRule.java
@@ -1,4 +1,4 @@
-package domain.move.moverule;
+package domain.move.rule;
 
 import domain.board.Intersection;
 import domain.game.Side;
@@ -6,7 +6,7 @@ import domain.piece.AlivePieces;
 import domain.move.movement.Path;
 import java.util.List;
 
-public final class BasicMoveRule {
+public final class BasicRule implements MoveRule {
 
     // TODO 필터링 메서드를 유틸로 분리해서 가독성 좋게 네이밍 고민
     public List<Intersection> movableDestinations(Side side, List<Path> candidatePaths, AlivePieces alivePieces) {
